refactor: move contrastive loss code to util.py
